FIXUP: CHROMIUM: remoteproc: mediatek: Add APU config and timesync
Fix warning:

   drivers/remoteproc/mediatek_apusys/mtk_apu_config.c: In function 'mtk_apu_config_setup':
>> drivers/remoteproc/mediatek_apusys/mtk_apu_config.c:120:26: error: assignment to 'uint64_t' {aka 'long long unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
     120 |         mdla_rv_mem->buf = mdla_buf;
         |                          ^

drivers/remoteproc/mediatek_apusys/plat/mt8196_plat.c:105:5: warning: no previous prototype for function 'apu_infra_lock' [-Wmissing-prototypes]
     105 | int apu_infra_lock(struct mtk_apu *apu, uint32_t op, enum apu_infra_bit_id id)
         |     ^

UPSTREAM-TASK=b:379040514
BUG=b:365876446
TEST=emerge-rauru chromeos-kernel-6_6

Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202501250531.TO9uLA61-lkp@intel.com/
Closes: https://lore.kernel.org/oe-kbuild-all/202501240446.RZdCv3pp-lkp@intel.com/
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6224195

+++ drivers/remoteproc/mediatek_apusys/mtk_apu_config.c
@@ -124,7 +124,7 @@ int mtk_apu_config_setup(struct mtk_apu *apu)
 		return -ENOMEM;
 	}
 	mdla_rv_mem->size = MDLA_RESERVE_MEM_SZ;
-	mdla_rv_mem->buf = mdla_buf;
+	mdla_rv_mem->buf = (uint64_t)mdla_buf;
 	mdla_rv_mem->da = mdla_da;
 
 	return 0;
